+++ tests/atest/transformers/AlignTestCases/test_transformer.py
@@ -1,9 +1,9 @@
 import pytest
 
-from .. import run_tidy_and_compare
+from .. import TransformerAcceptanceTest
 
 
-class TestAlignTestCases:
+class TestAlignTestCases(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "AlignTestCases"
 
     @pytest.mark.parametrize(
@@ -20,11 +20,10 @@ class TestAlignTestCases:
         ],
     )
     def test_transformer(self, source):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source=source, expected=source)
+        self.compare(source=source, expected=source)
 
     def test_only_with_headers(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="no_header_col.robot",
             expected="no_header_col_only_headers.robot",
             config=":only_with_headers=True",