+++ tests/atest/transformers/RenameKeywords/test_transformer.py
@@ -1,39 +1,35 @@
-from .. import run_tidy_and_compare, run_tidy
+from .. import TransformerAcceptanceTest
 
 
-class TestRenameKeywords:
+class TestRenameKeywords(TransformerAcceptanceTest):
     TRANSFORMER_NAME = "RenameKeywords"
 
     def test_transformer(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="test.robot", expected="test.robot")
+        self.compare(source="test.robot", expected="test.robot")
 
     def test_renaming_pattern(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="rename_pattern_partial.robot",
             config=r":replace_pattern=(?i)rename\s?me:replace_to=New_Shining_Name",
         )
 
     def test_renaming_whole_name_pattern(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="rename_pattern_whole.robot",
             config=r":replace_pattern=(?i)^rename\s?me$:replace_to=New_Shining_Name",
         )
 
     def test_keep_underscores(self):
-        run_tidy_and_compare(
-            self.TRANSFORMER_NAME,
+        self.compare(
             source="test.robot",
             expected="with_underscores.robot",
             config=r":remove_underscores=False",
         )
 
     def test_invalid_pattern(self):
-        result = run_tidy(
-            self.TRANSFORMER_NAME,
+        result = self.run_tidy(
             args=rf"--transform {self.TRANSFORMER_NAME}:replace_pattern=[\911]".split(),
             source="test.robot",
             exit_code=1,
@@ -47,4 +43,4 @@ class TestRenameKeywords:
         assert expected_output in str(result.exception)
 
     def test_with_library_name(self):
-        run_tidy_and_compare(self.TRANSFORMER_NAME, source="with_library_name.robot")
+        self.compare(source="with_library_name.robot")